huggingface/requirements.txt
torch==2.5.1
torchvision==0.20.1
gradio>=4.0.0
huggingface_hub==0.26.2
pillow>=10.0.0
requests>=2.31.0
